Refactor query rewriter to interfaces and implementations to allow customization
