+++ pinot-common/src/main/java/org/apache/pinot/sql/parsers/CalciteSqlParser.java
@@ -20,7 +20,6 @@ package org.apache.pinot.sql.parsers;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -29,7 +28,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import javax.annotation.Nullable;
 import org.apache.calcite.config.Lex;
 import org.apache.calcite.sql.SqlBasicCall;
 import org.apache.calcite.sql.SqlDataTypeSpec;
@@ -50,19 +48,15 @@ import org.apache.calcite.sql.validate.SqlConformanceEnum;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.pinot.common.function.FunctionDefinitionRegistry;
-import org.apache.pinot.common.function.FunctionInfo;
-import org.apache.pinot.common.function.FunctionInvoker;
-import org.apache.pinot.common.function.FunctionRegistry;
-import org.apache.pinot.common.function.TransformFunctionType;
 import org.apache.pinot.common.request.DataSource;
 import org.apache.pinot.common.request.Expression;
 import org.apache.pinot.common.request.ExpressionType;
 import org.apache.pinot.common.request.Function;
-import org.apache.pinot.common.request.Identifier;
 import org.apache.pinot.common.request.PinotQuery;
 import org.apache.pinot.common.utils.request.RequestUtils;
-import org.apache.pinot.pql.parsers.pql2.ast.FilterKind;
 import org.apache.pinot.segment.spi.AggregationFunctionType;
+import org.apache.pinot.sql.parsers.rewriter.QueryRewriter;
+import org.apache.pinot.sql.parsers.rewriter.QueryRewriterFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -85,6 +79,8 @@ public class CalciteSqlParser {
   private static final SqlParser.Config PARSER_CONFIG =
       SqlParser.configBuilder().setLex(PINOT_LEX).setConformance(SqlConformanceEnum.BABEL)
           .setParserFactory(SqlBabelParserImpl.FACTORY).build();
+
+  public static final List<QueryRewriter> QUERY_REWRITERS = new ArrayList<>(QueryRewriterFactory.getQueryRewriters());
 
   // To Keep the backward compatibility with 'OPTION' Functionality in PQL, which is used to
   // provide more hints for query processing.
@@ -114,48 +110,10 @@ public class CalciteSqlParser {
     return pinotQuery;
   }
 
-  static void validate(Map<Identifier, Expression> aliasMap, PinotQuery pinotQuery)
+  static void validate(PinotQuery pinotQuery)
       throws SqlCompilationException {
-    validateSelectionClause(aliasMap, pinotQuery);
     validateGroupByClause(pinotQuery);
     validateDistinctQuery(pinotQuery);
-  }
-
-  private static void validateSelectionClause(Map<Identifier, Expression> aliasMap, PinotQuery pinotQuery)
-      throws SqlCompilationException {
-    // Sanity check on selection expression shouldn't use alias reference.
-    Set<String> aliasKeys = new HashSet<>();
-    for (Identifier identifier : aliasMap.keySet()) {
-      String aliasName = identifier.getName().toLowerCase();
-      if (!aliasKeys.add(aliasName)) {
-        throw new SqlCompilationException("Duplicated alias name found.");
-      }
-    }
-    for (Expression selectExpr : pinotQuery.getSelectList()) {
-      matchIdentifierInAliasMap(selectExpr, aliasKeys);
-    }
-  }
-
-  private static void matchIdentifierInAliasMap(Expression selectExpr, Set<String> aliasKeys)
-      throws SqlCompilationException {
-    Function functionCall = selectExpr.getFunctionCall();
-    if (functionCall != null) {
-      if (functionCall.getOperator().equalsIgnoreCase(SqlKind.AS.toString())) {
-        matchIdentifierInAliasMap(functionCall.getOperands().get(0), aliasKeys);
-      } else {
-        if (functionCall.getOperandsSize() > 0) {
-          for (Expression operand : functionCall.getOperands()) {
-            matchIdentifierInAliasMap(operand, aliasKeys);
-          }
-        }
-      }
-    }
-    if (selectExpr.getIdentifier() != null) {
-      if (aliasKeys.contains(selectExpr.getIdentifier().getName().toLowerCase())) {
-        throw new SqlCompilationException(
-            "Alias " + selectExpr.getIdentifier().getName() + " cannot be referred in SELECT Clause");
-      }
-    }
   }
 
   private static void validateGroupByClause(PinotQuery pinotQuery)
@@ -237,7 +195,7 @@ public class CalciteSqlParser {
     return true;
   }
 
-  private static boolean isAggregateExpression(Expression expression) {
+  public static boolean isAggregateExpression(Expression expression) {
     Function functionCall = expression.getFunctionCall();
     if (functionCall != null) {
       String operator = functionCall.getOperator();
@@ -257,6 +215,10 @@ public class CalciteSqlParser {
     return false;
   }
 
+  public static boolean isAsFunction(Expression expression) {
+    return expression.getFunctionCall() != null && expression.getFunctionCall().getOperator().equalsIgnoreCase("AS");
+  }
+
   /**
    * Extract all the identifiers from given expressions.
    *
@@ -271,8 +233,8 @@ public class CalciteSqlParser {
         identifiers.add(expression.getIdentifier().getName());
       } else if (expression.getFunctionCall() != null) {
         if (excludeAs && expression.getFunctionCall().getOperator().equalsIgnoreCase("AS")) {
-          identifiers
-              .addAll(extractIdentifiers(Arrays.asList(expression.getFunctionCall().getOperands().get(0)), true));
+          identifiers.addAll(
+              extractIdentifiers(Arrays.asList(expression.getFunctionCall().getOperands().get(0)), true));
           continue;
         } else {
           identifiers.addAll(extractIdentifiers(expression.getFunctionCall().getOperands(), excludeAs));
@@ -394,352 +356,11 @@ public class CalciteSqlParser {
   }
 
   private static void queryRewrite(PinotQuery pinotQuery) {
-    // Invoke compilation time functions
-    invokeCompileTimeFunctions(pinotQuery);
-
-    // Rewrite Selection list
-    rewriteSelections(pinotQuery.getSelectList());
-
-    // Update Predicate Comparison
-    Expression filterExpression = pinotQuery.getFilterExpression();
-    if (filterExpression != null) {
-      pinotQuery.setFilterExpression(updateComparisonPredicate(filterExpression));
+    for (QueryRewriter queryRewriter : QUERY_REWRITERS) {
+      pinotQuery = queryRewriter.rewrite(pinotQuery);
     }
-    Expression havingExpression = pinotQuery.getHavingExpression();
-    if (havingExpression != null) {
-      pinotQuery.setHavingExpression(updateComparisonPredicate(havingExpression));
-    }
-
-    // Update Ordinals
-    applyOrdinals(pinotQuery);
-
-    // Rewrite GroupBy to Distinct
-    rewriteNonAggregationGroupByToDistinct(pinotQuery);
-
-    // Update alias
-    Map<Identifier, Expression> aliasMap = extractAlias(pinotQuery.getSelectList());
-    applyAlias(aliasMap, pinotQuery);
-
     // Validate
-    validate(aliasMap, pinotQuery);
-  }
-
-  private static void applyOrdinals(PinotQuery pinotQuery) {
-    // handle GROUP BY clause
-    for (int i = 0; i < pinotQuery.getGroupByListSize(); i++) {
-      final Expression groupByExpr = pinotQuery.getGroupByList().get(i);
-      if (groupByExpr.isSetLiteral() && groupByExpr.getLiteral().isSetLongValue()) {
-        final int ordinal = (int) groupByExpr.getLiteral().getLongValue();
-        pinotQuery.getGroupByList().set(i, getExpressionFromOrdinal(pinotQuery.getSelectList(), ordinal));
-      }
-    }
-
-    // handle ORDER BY clause
-    for (int i = 0; i < pinotQuery.getOrderByListSize(); i++) {
-      final Expression orderByExpr = pinotQuery.getOrderByList().get(i).getFunctionCall().getOperands().get(0);
-      if (orderByExpr.isSetLiteral() && orderByExpr.getLiteral().isSetLongValue()) {
-        final int ordinal = (int) orderByExpr.getLiteral().getLongValue();
-        pinotQuery.getOrderByList().get(i).getFunctionCall()
-            .setOperands(Arrays.asList(getExpressionFromOrdinal(pinotQuery.getSelectList(), ordinal)));
-      }
-    }
-  }
-
-  private static Expression getExpressionFromOrdinal(List<Expression> selectList, int ordinal) {
-    if (ordinal > 0 && ordinal <= selectList.size()) {
-      final Expression expression = selectList.get(ordinal - 1);
-      // If the expression has AS, return the left operand.
-      if (expression.isSetFunctionCall() && expression.getFunctionCall().getOperator().equals(SqlKind.AS.name())) {
-        return expression.getFunctionCall().getOperands().get(0);
-      }
-      return expression;
-    } else {
-      throw new SqlCompilationException(
-          String.format("Expected Ordinal value to be between 1 and %d.", selectList.size()));
-    }
-  }
-
-  private static void rewriteSelections(List<Expression> selectList) {
-    for (Expression expression : selectList) {
-      // Rewrite aggregation
-      tryToRewriteArrayFunction(expression);
-    }
-  }
-
-  private static void tryToRewriteArrayFunction(Expression expression) {
-    if (!expression.isSetFunctionCall()) {
-      return;
-    }
-    Function functionCall = expression.getFunctionCall();
-    switch (canonicalize(functionCall.getOperator())) {
-      case "sum":
-        if (functionCall.getOperands().size() != 1) {
-          return;
-        }
-        if (functionCall.getOperands().get(0).isSetFunctionCall()) {
-          Function innerFunction = functionCall.getOperands().get(0).getFunctionCall();
-          if (isSameFunction(innerFunction.getOperator(), TransformFunctionType.ARRAYSUM.getName())) {
-            Function sumMvFunc = new Function(AggregationFunctionType.SUMMV.getName());
-            sumMvFunc.setOperands(innerFunction.getOperands());
-            expression.setFunctionCall(sumMvFunc);
-          }
-        }
-        return;
-      case "min":
-        if (functionCall.getOperands().size() != 1) {
-          return;
-        }
-        if (functionCall.getOperands().get(0).isSetFunctionCall()) {
-          Function innerFunction = functionCall.getOperands().get(0).getFunctionCall();
-          if (isSameFunction(innerFunction.getOperator(), TransformFunctionType.ARRAYMIN.getName())) {
-            Function sumMvFunc = new Function(AggregationFunctionType.MINMV.getName());
-            sumMvFunc.setOperands(innerFunction.getOperands());
-            expression.setFunctionCall(sumMvFunc);
-          }
-        }
-        return;
-      case "max":
-        if (functionCall.getOperands().size() != 1) {
-          return;
-        }
-        if (functionCall.getOperands().get(0).isSetFunctionCall()) {
-          Function innerFunction = functionCall.getOperands().get(0).getFunctionCall();
-          if (isSameFunction(innerFunction.getOperator(), TransformFunctionType.ARRAYMAX.getName())) {
-            Function sumMvFunc = new Function(AggregationFunctionType.MAXMV.getName());
-            sumMvFunc.setOperands(innerFunction.getOperands());
-            expression.setFunctionCall(sumMvFunc);
-          }
-        }
-        return;
-      default:
-        break;
-    }
-    for (Expression operand : functionCall.getOperands()) {
-      tryToRewriteArrayFunction(operand);
-    }
-  }
-
-  private static String canonicalize(String functionName) {
-    return StringUtils.remove(functionName, '_').toLowerCase();
-  }
-
-  private static boolean isSameFunction(String function1, String function2) {
-    return canonicalize(function1).equals(canonicalize(function2));
-  }
-
-  /**
-   * Rewrite non-aggregate group by query to distinct query.
-   * E.g.
-   * ```
-   *   SELECT col1+col2*5 FROM foo GROUP BY col1, col2 => SELECT distinct col1+col2*5 FROM foo
-   *   SELECT col1, col2 FROM foo GROUP BY col1, col2 => SELECT distinct col1, col2 FROM foo
-   * ```
-   * @param pinotQuery
-   */
-  private static void rewriteNonAggregationGroupByToDistinct(PinotQuery pinotQuery) {
-    boolean hasAggregation = false;
-    for (Expression select : pinotQuery.getSelectList()) {
-      if (isAggregateExpression(select)) {
-        hasAggregation = true;
-      }
-    }
-    if (pinotQuery.getOrderByList() != null) {
-      for (Expression orderBy : pinotQuery.getOrderByList()) {
-        if (isAggregateExpression(orderBy)) {
-          hasAggregation = true;
-        }
-      }
-    }
-    if (!hasAggregation && pinotQuery.getGroupByListSize() > 0) {
-      Set<String> selectIdentifiers = extractIdentifiers(pinotQuery.getSelectList(), true);
-      Set<String> groupByIdentifiers = extractIdentifiers(pinotQuery.getGroupByList(), true);
-      if (groupByIdentifiers.containsAll(selectIdentifiers)) {
-        Expression distinctExpression = RequestUtils.getFunctionExpression("DISTINCT");
-        for (Expression select : pinotQuery.getSelectList()) {
-          if (isAsFunction(select)) {
-            Function asFunc = select.getFunctionCall();
-            distinctExpression.getFunctionCall().addToOperands(asFunc.getOperands().get(0));
-          } else {
-            distinctExpression.getFunctionCall().addToOperands(select);
-          }
-        }
-        pinotQuery.setSelectList(Arrays.asList(distinctExpression));
-        pinotQuery.setGroupByList(Collections.emptyList());
-      } else {
-        selectIdentifiers.removeAll(groupByIdentifiers);
-        throw new SqlCompilationException(String.format(
-            "For non-aggregation group by query, all the identifiers in select clause should be in groupBys. Found "
-                + "identifier: %s", Arrays.toString(selectIdentifiers.toArray(new String[0]))));
-      }
-    }
-  }
-
-  private static boolean isAsFunction(Expression expression) {
-    return expression.getFunctionCall() != null && expression.getFunctionCall().getOperator().equalsIgnoreCase("AS");
-  }
-
-  private static void invokeCompileTimeFunctions(PinotQuery pinotQuery) {
-    for (int i = 0; i < pinotQuery.getSelectListSize(); i++) {
-      Expression expression = invokeCompileTimeFunctionExpression(pinotQuery.getSelectList().get(i));
-      pinotQuery.getSelectList().set(i, expression);
-    }
-    for (int i = 0; i < pinotQuery.getGroupByListSize(); i++) {
-      Expression expression = invokeCompileTimeFunctionExpression(pinotQuery.getGroupByList().get(i));
-      pinotQuery.getGroupByList().set(i, expression);
-    }
-    for (int i = 0; i < pinotQuery.getOrderByListSize(); i++) {
-      Expression expression = invokeCompileTimeFunctionExpression(pinotQuery.getOrderByList().get(i));
-      pinotQuery.getOrderByList().set(i, expression);
-    }
-    Expression filterExpression = invokeCompileTimeFunctionExpression(pinotQuery.getFilterExpression());
-    pinotQuery.setFilterExpression(filterExpression);
-    Expression havingExpression = invokeCompileTimeFunctionExpression(pinotQuery.getHavingExpression());
-    pinotQuery.setHavingExpression(havingExpression);
-  }
-
-  // This method converts a predicate expression to the what Pinot could evaluate.
-  // For comparison expression, left operand could be any expression, but right operand only
-  // supports literal.
-  // E.g. 'WHERE a > b' will be updated to 'WHERE a - b > 0'
-  private static Expression updateComparisonPredicate(Expression expression) {
-    Function function = expression.getFunctionCall();
-    if (function != null) {
-      String operator = function.getOperator().toUpperCase();
-      FilterKind filterKind;
-      try {
-        filterKind = FilterKind.valueOf(operator);
-      } catch (Exception e) {
-        throw new SqlCompilationException("Unsupported filter kind: " + operator);
-      }
-      List<Expression> operands = function.getOperands();
-      switch (filterKind) {
-        case AND:
-        case OR:
-          operands.replaceAll(CalciteSqlParser::updateComparisonPredicate);
-          break;
-        case EQUALS:
-        case NOT_EQUALS:
-        case GREATER_THAN:
-        case GREATER_THAN_OR_EQUAL:
-        case LESS_THAN:
-        case LESS_THAN_OR_EQUAL:
-          Expression firstOperand = operands.get(0);
-          Expression secondOperand = operands.get(1);
-
-          // Handle predicate like '10 = a' -> 'a = 10'
-          if (firstOperand.isSetLiteral()) {
-            if (!secondOperand.isSetLiteral()) {
-              function.setOperator(getOppositeOperator(filterKind).name());
-              operands.set(0, secondOperand);
-              operands.set(1, firstOperand);
-            }
-            break;
-          }
-
-          // Handle predicate like 'a > b' -> 'a - b > 0'
-          if (!secondOperand.isSetLiteral()) {
-            Expression minusExpression = RequestUtils.getFunctionExpression(SqlKind.MINUS.name());
-            minusExpression.getFunctionCall().setOperands(Arrays.asList(firstOperand, secondOperand));
-            operands.set(0, minusExpression);
-            operands.set(1, RequestUtils.getLiteralExpression(0));
-            break;
-          }
-          break;
-        default:
-          int numOperands = operands.size();
-          for (int i = 1; i < numOperands; i++) {
-            if (!operands.get(i).isSetLiteral()) {
-              throw new SqlCompilationException(String
-                  .format("For %s predicate, the operands except for the first one must be literal, got: %s",
-                      filterKind, expression));
-            }
-          }
-          break;
-      }
-    }
-    return expression;
-  }
-
-  /**
-   * The purpose of this method is to convert expression "0 < columnA" to "columnA > 0".
-   * The conversion would be:
-   *  from ">" to "<",
-   *  from "<" to ">",
-   *  from ">=" to "<=",
-   *  from "<=" to ">=".
-   */
-  private static FilterKind getOppositeOperator(FilterKind filterKind) {
-    switch (filterKind) {
-      case GREATER_THAN:
-        return FilterKind.LESS_THAN;
-      case GREATER_THAN_OR_EQUAL:
-        return FilterKind.LESS_THAN_OR_EQUAL;
-      case LESS_THAN:
-        return FilterKind.GREATER_THAN;
-      case LESS_THAN_OR_EQUAL:
-        return FilterKind.GREATER_THAN_OR_EQUAL;
-      default:
-        // Do nothing
-        return filterKind;
-    }
-  }
-
-  private static void applyAlias(Map<Identifier, Expression> aliasMap, PinotQuery pinotQuery) {
-    Expression filterExpression = pinotQuery.getFilterExpression();
-    if (filterExpression != null) {
-      applyAlias(aliasMap, filterExpression);
-    }
-    List<Expression> groupByList = pinotQuery.getGroupByList();
-    if (groupByList != null) {
-      for (Expression expression : groupByList) {
-        applyAlias(aliasMap, expression);
-      }
-    }
-    Expression havingExpression = pinotQuery.getHavingExpression();
-    if (havingExpression != null) {
-      applyAlias(aliasMap, havingExpression);
-    }
-    List<Expression> orderByList = pinotQuery.getOrderByList();
-    if (orderByList != null) {
-      for (Expression expression : orderByList) {
-        applyAlias(aliasMap, expression);
-      }
-    }
-  }
-
-  private static void applyAlias(Map<Identifier, Expression> aliasMap, Expression expression) {
-    Identifier identifierKey = expression.getIdentifier();
-    if (identifierKey != null) {
-      Expression aliasExpression = aliasMap.get(identifierKey);
-      if (aliasExpression != null) {
-        expression.setType(aliasExpression.getType());
-        expression.setIdentifier(aliasExpression.getIdentifier());
-        expression.setFunctionCall(aliasExpression.getFunctionCall());
-        expression.setLiteral(aliasExpression.getLiteral());
-      }
-      return;
-    }
-    Function function = expression.getFunctionCall();
-    if (function != null) {
-      for (Expression operand : function.getOperands()) {
-        applyAlias(aliasMap, operand);
-      }
-    }
-  }
-
-  private static Map<Identifier, Expression> extractAlias(List<Expression> expressions) {
-    Map<Identifier, Expression> aliasMap = new HashMap<>();
-    for (Expression expression : expressions) {
-      Function functionCall = expression.getFunctionCall();
-      if (functionCall == null) {
-        continue;
-      }
-      if (functionCall.getOperator().equalsIgnoreCase(SqlKind.AS.toString())) {
-        Expression identifierExpr = functionCall.getOperands().get(1);
-        aliasMap.put(identifierExpr.getIdentifier(), functionCall.getOperands().get(0));
-      }
-    }
-    return aliasMap;
+    validate(pinotQuery);
   }
 
   private static List<String> extractOptionsFromSql(String sql) {
@@ -991,7 +612,7 @@ public class CalciteSqlParser {
     // Compile first operand of the function (either an identifier or another DOT and/or ITEM function).
     SqlKind kind0 = operands[0].getKind();
     if (kind0 == SqlKind.IDENTIFIER) {
-      path.append(((SqlIdentifier) operands[0]).toString());
+      path.append(operands[0].toString());
     } else if (kind0 == SqlKind.DOT || kind0 == SqlKind.OTHER_FUNCTION) {
       SqlBasicCall function0 = (SqlBasicCall) operands[0];
       String name0 = function0.getOperator().getName();
@@ -1013,6 +634,15 @@ public class CalciteSqlParser {
     } else {
       throw new SqlCompilationException("SELECT list item has bad path expression.");
     }
+  }
+
+
+  public static String canonicalize(String functionName) {
+    return StringUtils.remove(functionName, '_').toLowerCase();
+  }
+
+  public static boolean isSameFunction(String function1, String function2) {
+    return canonicalize(function1).equals(canonicalize(function2));
   }
 
   private static void validateFunction(String functionName, List<Expression> operands) {
@@ -1092,44 +722,6 @@ public class CalciteSqlParser {
     Expression andExpression = RequestUtils.getFunctionExpression(SqlKind.OR.name());
     andExpression.getFunctionCall().setOperands(operands);
     return andExpression;
-  }
-
-  protected static Expression invokeCompileTimeFunctionExpression(@Nullable Expression expression) {
-    if (expression == null || expression.getFunctionCall() == null) {
-      return expression;
-    }
-    Function function = expression.getFunctionCall();
-    List<Expression> operands = function.getOperands();
-    int numOperands = operands.size();
-    boolean compilable = true;
-    for (int i = 0; i < numOperands; i++) {
-      Expression operand = invokeCompileTimeFunctionExpression(operands.get(i));
-      if (operand.getLiteral() == null) {
-        compilable = false;
-      }
-      operands.set(i, operand);
-    }
-    String functionName = function.getOperator();
-    if (compilable) {
-      FunctionInfo functionInfo = FunctionRegistry.getFunctionInfo(functionName, numOperands);
-      if (functionInfo != null) {
-        Object[] arguments = new Object[numOperands];
-        for (int i = 0; i < numOperands; i++) {
-          arguments[i] = function.getOperands().get(i).getLiteral().getFieldValue();
-        }
-        try {
-          FunctionInvoker invoker = new FunctionInvoker(functionInfo);
-          invoker.convertTypes(arguments);
-          Object result = invoker.invoke(arguments);
-          return RequestUtils.getLiteralExpression(result);
-        } catch (Exception e) {
-          throw new SqlCompilationException(
-              "Caught exception while invoking method: " + functionInfo.getMethod() + " with arguments: " + Arrays
-                  .toString(arguments), e);
-        }
-      }
-    }
-    return expression;
   }
 
   public static boolean isLiteralOnlyExpression(Expression e) {
