+++ pinot-common/src/test/java/org/apache/pinot/sql/parsers/CalciteSqlCompilerTest.java
@@ -40,6 +40,7 @@ import org.apache.pinot.common.utils.request.FilterQueryTree;
 import org.apache.pinot.common.utils.request.RequestUtils;
 import org.apache.pinot.pql.parsers.PinotQuery2BrokerRequestConverter;
 import org.apache.pinot.segment.spi.AggregationFunctionType;
+import org.apache.pinot.sql.parsers.rewriter.CompileTimeFunctionsInvoker;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -1767,10 +1768,14 @@ public class CalciteSqlCompilerTest {
 
   @Test
   public void testCompileTimeExpression() {
+    final CompileTimeFunctionsInvoker compileTimeFunctionsInvoker = new CompileTimeFunctionsInvoker();
     long lowerBound = System.currentTimeMillis();
     Expression expression = CalciteSqlParser.compileToExpression("now()");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    PinotQuery pinotQuery = new PinotQuery();
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getLiteral());
     long upperBound = System.currentTimeMillis();
     long result = expression.getLiteral().getLongValue();
@@ -1779,7 +1784,9 @@ public class CalciteSqlCompilerTest {
     lowerBound = TimeUnit.MILLISECONDS.toHours(System.currentTimeMillis()) + 1;
     expression = CalciteSqlParser.compileToExpression("to_epoch_hours(now() + 3600000)");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getLiteral());
     upperBound = TimeUnit.MILLISECONDS.toHours(System.currentTimeMillis()) + 1;
     result = expression.getLiteral().getLongValue();
@@ -1788,7 +1795,9 @@ public class CalciteSqlCompilerTest {
     lowerBound = System.currentTimeMillis() - ONE_HOUR_IN_MS;
     expression = CalciteSqlParser.compileToExpression("ago('PT1H')");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getLiteral());
     upperBound = System.currentTimeMillis() - ONE_HOUR_IN_MS;
     result = expression.getLiteral().getLongValue();
@@ -1797,7 +1806,9 @@ public class CalciteSqlCompilerTest {
     lowerBound = System.currentTimeMillis() + ONE_HOUR_IN_MS;
     expression = CalciteSqlParser.compileToExpression("ago('PT-1H')");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getLiteral());
     upperBound = System.currentTimeMillis() + ONE_HOUR_IN_MS;
     result = expression.getLiteral().getLongValue();
@@ -1805,7 +1816,9 @@ public class CalciteSqlCompilerTest {
 
     expression = CalciteSqlParser.compileToExpression("toDateTime(millisSinceEpoch)");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getFunctionCall());
     Assert.assertEquals(expression.getFunctionCall().getOperator(), "TODATETIME");
     Assert
@@ -1813,26 +1826,34 @@ public class CalciteSqlCompilerTest {
 
     expression = CalciteSqlParser.compileToExpression("reverse(playerName)");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getFunctionCall());
     Assert.assertEquals(expression.getFunctionCall().getOperator(), "REVERSE");
     Assert.assertEquals(expression.getFunctionCall().getOperands().get(0).getIdentifier().getName(), "playerName");
 
     expression = CalciteSqlParser.compileToExpression("reverse('playerName')");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getLiteral());
     Assert.assertEquals(expression.getLiteral().getFieldValue(), "emaNreyalp");
 
     expression = CalciteSqlParser.compileToExpression("reverse(123)");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getLiteral());
     Assert.assertEquals(expression.getLiteral().getFieldValue(), "321");
 
     expression = CalciteSqlParser.compileToExpression("count(*)");
     Assert.assertNotNull(expression.getFunctionCall());
-    expression = CalciteSqlParser.invokeCompileTimeFunctionExpression(expression);
+    pinotQuery.setFilterExpression(expression);
+    pinotQuery = compileTimeFunctionsInvoker.rewrite(pinotQuery);
+    expression = pinotQuery.getFilterExpression();
     Assert.assertNotNull(expression.getFunctionCall());
     Assert.assertEquals(expression.getFunctionCall().getOperator(), "COUNT");
     Assert.assertEquals(expression.getFunctionCall().getOperands().get(0).getIdentifier().getName(), "*");
